UPD: Refactor game rooms
